% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy64-package.R
\docType{package}
\name{tidy64-package}
\alias{tidy64}
\alias{tidy64-package}
\title{tidy64: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
