% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{tidy64-vctrs-compat}
\alias{tidy64-vctrs-compat}
\alias{vec_arith.tidy64}
\title{vctrs compatibility functions}
\usage{
\method{vec_arith}{tidy64}(op, x, y, ...)
}
\arguments{
\item{op}{An arithmetic operator as a string.}

\item{x, y}{Objects.}

\item{...}{Used to pass along error message information.}
}
\value{
See the corresponding vctrs function for the exact return value.
}
\description{
These functions are the extensions that allow stepper objects to
work with vctrs.
}
