% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy64-package.R
\docType{package}
\name{tidy64-package}
\alias{tidy64-package}
\alias{_PACKAGE}
\title{tidy64: An S3 Class Supporting 64-bit Integers}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/tidy64}
  \item Report bugs at \url{https://github.com/DavisVaughan/tidy64/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
